requirements.txt
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Pydantic / settings
pydantic>=2.9.0
pydantic-settings>=2.5.0

# HTTP client (used by Tavily service — NOT for Google auth)
httpx>=0.27.0

# Utilities
python-dateutil>=2.9.0
cachetools>=5.5.0
python-dotenv>=1.0.0

# ── Auth ──────────────────────────────────────────────────────────────────────
PyJWT>=2.8.0          # Create/verify our own JWTs
google-auth>=2.28.0   # Verify Google ID tokens locally (no API calls)
requests>=2.31.0      # Required by google-auth transport layer
